# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Group Internal Procurement MVP API
  version: 0.3.0
  description: >
    Formal MVP API contract covering P0 controls and Phase 1 supplier management,
    procurement request and project initiation. This contract keeps the MVP boundary:
    no CA, e-signature, bid-file encryption/decryption, trusted timestamp,
    tamper-proof evidence, opening hall, or real external system integration.
servers:
  - url: http://127.0.0.1:3000
security:
  - MockUserHeader: []
components:
  securitySchemes:
    MockUserHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            auditLogId:
              type: string
  responses:
    Ok:
      description: Request succeeded.
...
    PermissionDenied:
      description: Request rejected by backend policy and recorded with auditLogId when applicable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /api/roles:
    get:
      tags: [Organization]
      summary: Admin-only role configuration
      description: Only system administrators can read role configuration. Business, supplier, expert and audit users receive 403.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/role-permissions:
    get:
      tags: [Organization]
      summary: Admin-only role permission configuration
      description: Only system administrators can read role permission configuration. This endpoint is not a business authorization source for ordinary users.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/system-dictionaries:
    get:
      tags: [Organization]
      summary: Admin-only system dictionaries
      description: Only system administrators can read dictionary configuration. Procurement method lookup uses /api/procurement-method-rules instead.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /api/procurement-method-rules:
    get:
      tags: [Procurement Requests]
      summary: Procurement method rules visible to procurement/audit/config roles
      description: Buyer, group manager, auditor and admin can read method rules. Supplier and expert users receive 403 so procurement policy configuration is not exposed through portal roles.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...